requirements.txt
fastapi==0.110.0
python_jose==3.3.0
numpy==1.24.3
passlib==1.7.4
pydantic==2.6.4
pytest==8.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.23
torch==2.2.2
transformers==4.39.0
uvicorn==0.28.0
python-multipart==0.0.7
email-validator==2.1.1
torchaudio==2.2.2
torchvision==0.17.2
psycopg2-binary



